requirements.txt
APScheduler==3.8.1
bcrypt==3.2.0
black==21.10b0
blinker==1.4
cffi==1.15.0
cfgv==3.3.1
click==8.0.3
distlib==0.3.4
filelock==3.4.2
flake8==4.0.1
Flask==2.0.2
Flask-APScheduler==1.12.3
Flask-Cors==3.0.10
Flask-JWT-Extended==4.3.1
Flask-Mail==0.9.1
gevent==21.12.0
greenlet==1.1.2
gunicorn==20.1.0
identify==2.4.4
install==1.3.5
itsdangerous==2.0.1
Jinja2==3.0.2
ldap3==2.9.1
MarkupSafe==2.0.1
mccabe==0.6.1
mypy-extensions==0.4.3
mysql-connector-python==8.0.27
nodeenv==1.6.0
pathspec==0.9.0
Pillow==9.0.1
platformdirs==2.4.0
pre-commit==2.15.0
protobuf==3.19.1
pyasn1==0.4.8
pycodestyle==2.8.0
pycparser==2.21
pyflakes==2.4.0
PyJWT==2.3.0
python-dateutil==2.8.2
python-dotenv==0.19.1
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
regex==2021.11.1
six==1.16.0
toml==0.10.2
tomli==1.2.2
typing-extensions==3.10.0.2
tzdata==2021.5
tzlocal==4.1
virtualenv==20.13.0
Werkzeug==2.0.2
zope.event==4.5.0
zope.interface==5.4.0
